Avoid monomorphizing encoders if not used (#2468)
