+++ src/buffer.rs
@@ -7,7 +7,6 @@ use std::path::Path;
 use std::slice::{ChunksExact, ChunksExactMut};
 
 use crate::color::{FromColor, Luma, LumaA, Rgb, Rgba};
-use crate::dynimage::{save_buffer, save_buffer_with_format, write_buffer_with_format};
 use crate::error::ImageResult;
 use crate::flat::{FlatSamples, SampleLayout};
 use crate::image::{GenericImage, GenericImageView, ImageEncoder, ImageFormat};
@@ -15,6 +14,7 @@ use crate::math::Rect;
 use crate::traits::{EncodableLayout, Pixel, PixelWithColorType};
 use crate::utils::expand_packed;
 use crate::DynamicImage;
+use crate::{save_buffer, save_buffer_with_format, write_buffer_with_format};
 
 /// Iterate over pixel refs.
 pub struct Pixels<'a, P: Pixel + 'a>
